fix: preseed script

+++ src/preseed.sh
@@ -1,21 +1,19 @@
 #!/bin/sh
-set -eu
+set -e
 
+if [ "${1:-}" = "--worker" ]; then
+    line="${2?}"
+    hash="${line%% *}"
+    package="${line#* }"
+    ref="stagex/${package}@sha256:${hash}"
+    docker pull "${ref}"
+    rm -rf "out/${package}"
+    mkdir -p "out/${package}"
+    docker save "${ref}" | tar -xC "out/${package}"
+    exit 0
+fi
+
+set -u
 date="$(date '+%F %T')"
-
-fetch(){
-	line=${1?}
-	local package=$(echo ${line} | sed 's/^[a-z0-9]\+ \(.*\)/\1/g');
-	local hash=$(echo ${line} | sed 's/^\([a-z0-9]\+\) .*/\1/g');
-	local ref=$(printf "stagex/%s@sha256:%s" "$package" "$hash");
-	docker pull ${ref}
-	rm -rf "out/${package}"
-	mkdir -p "out/${package}"
-	docker save ${ref} | tar -xC out/${package}
-}
-
-cat digests/* | while read line; do
-	fetch "${line}"
-done
-
+cat digests/* | xargs -I {} -P $(nproc) "$0" --worker "{}"
 find out -type f -exec touch -d "$date" {} +
